* Rearranged some constructor initializations to prevent compiler warnings. * Changed powf() to pow() to work around a mingw runtime bug. See http://bugs.debian.org/484299 * Changed abs(float) to fabs(float). abs() only has an int and long version, and mingw was complaining about ambiguity. (float is what we wanted anyway). * Commented out a couple of luabind functions that were preventing mingw from compiling. According to the comments, they gave gcc problems too, and were not used.

+++ demo/src/engine/video/interpolator.cpp
@@ -143,13 +143,13 @@ void Interpolator::Update(uint32 frame_time) {
 
 
 float Interpolator::_FastTransform(float initial_value) {
-	return powf(initial_value, VIDEO_FAST_TRANSFORM_POWER);
+	return pow(initial_value, VIDEO_FAST_TRANSFORM_POWER);
 }
 
 
 
 float Interpolator::_SlowTransform(float initial_value) {
-	return powf(initial_value, VIDEO_SLOW_TRANSFORM_POWER);
+	return pow(initial_value, VIDEO_SLOW_TRANSFORM_POWER);
 }
